+++ tools/importPlanGoals.test.js
@@ -1,4 +1,5 @@
 import importGoals from './importPlanGoals';
+import updateGrantsGrantees from '../src/lib/updateGrantsGrantees';
 import db, {
   Role, Topic, RoleTopic, Goal, Grantee, Grant, sequelize,
 } from '../src/models';
@@ -6,6 +7,39 @@ import db, {
 describe('Import TTA plan goals', () => {
   afterAll(() => {
     db.sequelize.close();
+  });
+  describe('Update grants and grantees', () => {
+    beforeAll(async () => {
+      await Grant.destroy({ where: {} });
+      await Grantee.destroy({ where: {} });
+    });
+    afterEach(async () => {
+      await Grant.destroy({ where: {} });
+      await Grantee.destroy({ where: {} });
+    });
+    it('should import or update grantees', async () => {
+
+      const granteesBefore = await Grantee.findAll();
+      expect(granteesBefore.length).toBe(0);
+      await updateGrantsGrantees();
+
+      const grantee = await Grantee.findOne({ where: { id: 1335 } });
+      expect(grantee).toBeDefined();
+      expect(grantee.name).toBe('Greater Bergen Community Action, Inc.');
+    });
+
+    it('should import or update grants', async () => {
+      const grantsBefore = await Grantee.findAll();
+
+      expect(grantsBefore.length).toBe(0);
+      await updateGrantsGrantees();
+
+      const grants = await Grant.findAll({ where: { granteeId: 1335 } });
+      expect(grants).toBeDefined();
+      expect(grants.length).toBe(3);
+      const containsNumber = grants.some((g) => g.number === '02CH010840');
+      expect(containsNumber).toBeTruthy();
+    });
   });
   it('should import Roles table', async () => {
     await Role.destroy({ where: {} });
